Add GenericComponentInteractionCreateEvent
